// src/app/model/EventText.ts
/**
 * Created by stijnergeerts on 15/05/17.
 */
export class EventText {
  eventName: string;
  logo: string;
  tabTitle: string;
  textUpper: string;
  textAbovePhoto: string;
  textUnderPhoto: string;
  textBottom: string;
  textColor: string;
  backgroundColor: string;
  backgroundImage: string;
  backgroundImageToggle: boolean;
  cloudinaryDirectory: string;
  qrCodeImage: string;
}